Add files via upload
Added 3x R programs. The overall functions of the code are described in the readme. 

Import data would be run first - it reads in the data for the four file types
DataWrangling then combines the four file types into the CompleteDataset
DataExplore performs the checks on data quality. Some issues discovered are indications of compliance with data collection standards and others with official guidance on child protection.